// src/app/core/models/cant-mes.ts
export interface CantMes {
    cantEnero: number;
    cantFebrero: number;
    cantMarzo: number;
    cantAbril: number;
    cantMayo: number;
    cantJunio: number;
    cantJulio: number;
    cantAgosto: number;
    cantSeptiembre: number;
    cantOctubre: number;
    cantNoviembre: number;
    cantDiciembre: number;
}